Crate sniffer_parser

Packet Parsing library from Ethernet frame to Application-layer representation

This library parses an Ethernet frame extracting all fields and data from it and represents the parsed packet data at the different levels of the TCP/IP stack

Modules

HeaderLength

Ethernet Header Length

dns

DNS Packet parsing

http

HTTP Packet parsing

serializable_packet

Generic parsed packet representation

tls

TLS Packet parsing

Enums

HttpPacketType

HTTP Types of packets

Functions

cleanup_sniffing_state

Delete active parsers

handle_application_protocol

Build an application-layer packet from a transport-layer one, save it in a Parsed Packet

handle_arp_packet

Build a ARP packet from a data-link packet, save it in a Parsed Packet

handle_icmp_packet

Build a ICMP packet from a network-layer packet, save it in a Parsed Packet

handle_icmpv6_packet

Build a ICMPv6 packet from a network-layer packet, save it in a Parsed Packet

handle_ipv4_packet

Build a IPv4 packet from a data-link packet, save it in a Parsed Packet

handle_ipv6_packet

Build a IPv6 packet from a data-link packet, save it in a Parsed Packet

handle_tcp_packet

Build a TCP packet from a network-layer packet, save it in a Parsed Packet

handle_transport_protocol

Build a Transport-layer packet from a network-layer packet, save it in a Parsed Packet

handle_udp_packet

Build a UDP packet from a network-layer packet, save it in a Parsed Packet

parse_ethernet_frame

Parse ethernet frame obtaining the packet link-layer and network-layer representations