Module sniffer_parser::serializable_packet

Generic parsed packet representation

A parsed packet contains the packet representation at each level of the TCP/IP stack (except physical):

link_layer_packet

network_layer_packet

transport_layer_packet

application_layer_packet

Modules

application

Application level Packets Representation

network

Network level Packets Representation

transport

Transport level Packets Representation

util

Utility functions to retrieve specific fields of packets

Structs

ParsedPacket

Data structure containing representations of the packet at each TCP/IP layer

SerializableEthernetPacket

Ethernet Packet Representation

SerializableUnknownPacket

Unknown Packet Representation

Enums

SerializablePacket

All possible packet serialization options